removed some fields, added simarchive outputs

+++ src/rebound.h
@@ -300,8 +300,8 @@ struct reb_integrator_trace {
     unsigned int current_C; // Tracking C for the entire timestep
     unsigned int force_accept; // Force accept for irreversible steps: collisions and adding particles
 
-    // For debugging
-    unsigned int step_rejections;
+    // For debugging, but maybe this is useful to have anyway?
+    uint64_t step_rejections;
 };
 
 // SABA Integrator (Laskar & Robutel 2001)
